Added rough asteroid collision physics
There's a new check for asteroid collisions with every other asteroid.
Only 1 collision between 2 bodies is accounted for, so strange things
happen when an asteroid collides with more than one other asteroid.
Collisions are assumed to be elastic. The collision check is currently
not very efficient, i.e. O(n^2)

Velocity of asteroids was changed to always be vectors instead of only
when updating the position.