Added config file support
The program looks for a configuration file called "asteroids.conf" in
the executable's directory. If the file does not exist, it is generated
using default options. If it exists, it is read and any recognized
options are used by the game. Any command line options take precedent
over config file options.

Also, asteroid collision physics are now on by default.