Added support for fullscreen and resolution
A user can now change the resolution in the 'asteroids.conf' file.
The values in the config file are interpreted as a target resolution
that is matched with the closest video mode available.

Screen bounds and projection now account for varying resolutions. These
are determined using SDL_GL_GetDrawableSize, which means this project
now requires at least SDL 2.0.1

A user can also change asteroid scale and mass modifiers with the new
command line switches and config file options.